Deep revisions + open findings: edge roles, run(), lazy structured graph, create_df, shortest paths, interactive_html split
Deep revisions:
- Model/colour inversion: new graphs/edge_roles.py (edge_role_of / base_color_of
  / EDGE_ROLE_*). tag_constrained_path reads the semantic role instead of parsing
  the Graphviz colour string; highlight_significant_line_loading records an
  authoritative base_color when it wraps a colour into the compound "c:yellow:c";
  new OverFlowGraph.edge_role(name). Exported via the package + shim.
- AlphaDeesp.__init__ -> explicit run(): the ranking pipeline moved into run();
  auto_run=True (default) preserves backwards-compatible construction; results
  are initialised empty so auto_run=False builds a well-formed, un-run object.
- Structured_Overload_Distribution_Graph: colour-filtered views, red_loops and
  hubs are now functools.cached_property (constrained path stays eager). red_loops
  uses the constructor seed hubs; public find_loops() uses the detected hubs — this
  split makes the lazy properties order-independent while matching the old eager
  behaviour exactly.

Open findings:
- #4 create_df: vectorised the iterrows passes with numpy masks; ltc_report now
  uses positional .iloc. Behaviour pinned by a 400-case differential fuzz against
  a faithful re-implementation of the original loop.
- #6 shortest_paths.py: shared _make_incentivized_weight, multigraph-correct
  (min parallel weight + (u,v)/(u,v,key) promoted matching); dead branch removed.
- #7 to_DiGraph: a missing capacity now defaults to 0.0 (neutral in the min-cut)
  instead of a spurious 1.0.

Maintainability:
- interactive_html.py (976 LOC) -> package interactive_html/ (8 focused modules)
  with the CSS/JS/HTML skeleton externalised under assets/ and reassembled
  byte-exactly at runtime; shipped via package_data + MANIFEST.in.

Docs: CLAUDE.md and docs/CODE_REVIEW.md updated. All graph-package / ranking /
renderer / create_df / interactive-html unit suites pass (367) without grid2op.
