+++ alphaDeesp/core/graphsAndPaths.py
@@ -11,6 +11,12 @@ from :mod:`alphaDeesp.core.graphs` (or one of its sub-modules).
 """
 
 from alphaDeesp.core.graphs import (  # noqa: F401
+    EDGE_ROLE_INSIGNIFICANT,
+    EDGE_ROLE_NEGATIVE,
+    EDGE_ROLE_NULL_NON_RECONNECTABLE,
+    EDGE_ROLE_OVERLOAD,
+    EDGE_ROLE_POSITIVE,
+    EDGE_ROLE_UNKNOWN,
     ConstrainedPath,
     OverFlowGraph,
     OverflowGraphRenderer,
@@ -18,8 +24,10 @@ from alphaDeesp.core.graphs import (  # noqa: F401
     Structured_Overload_Distribution_Graph,
     add_double_edges_null_redispatch,
     all_simple_edge_paths_multi,
+    base_color_of,
     default_voltage_colors,
     delete_color_edges,
+    edge_role_of,
     find_multidigraph_edges_by_name,
     from_edges_get_nodes,
     incident_edges,
@@ -37,6 +45,14 @@ __all__ = [
     "OverflowGraphRenderer",
     "ConstrainedPath",
     "Structured_Overload_Distribution_Graph",
+    "edge_role_of",
+    "base_color_of",
+    "EDGE_ROLE_OVERLOAD",
+    "EDGE_ROLE_NEGATIVE",
+    "EDGE_ROLE_POSITIVE",
+    "EDGE_ROLE_INSIGNIFICANT",
+    "EDGE_ROLE_NULL_NON_RECONNECTABLE",
+    "EDGE_ROLE_UNKNOWN",
     "from_edges_get_nodes",
     "delete_color_edges",
     "nodepath_to_edgepath",
