+++ alphaDeesp/core/simulation.py
@@ -8,7 +8,6 @@
 
 import logging
 from abc import ABC, abstractmethod
-from math import fabs
 from typing import Any, Dict, List, Optional, Sequence, Tuple, Union
 
 import numpy as np
@@ -211,7 +210,15 @@ class Simulation(ABC):
         return end_result_data_frame
 
     def create_df(self, d: Dict[str, Any], line_to_cut: List[int]) -> pd.DataFrame:
-        """arg: d represents a topology"""
+        """Build the topology + flow DataFrame for a cut line.
+
+        ``d`` represents a topology (one row per grid line, in line-id order).
+        The row-per-line ordering is a load-bearing invariant: the overloaded
+        line's delta flow is read positionally by ``line_to_cut`` id, and
+        ``OverFlowGraph`` matches ``lines_to_cut`` against row positions too.
+        The former ``iterrows`` passes are vectorised with numpy masks; the
+        numerical results are identical.
+        """
         # HERE WE CREATE DATAFRAME
         df = pd.DataFrame(d["edges"])
         pd.set_option("display.float_format", lambda x: "%.3f" % x)
@@ -219,78 +226,51 @@ class Simulation(ABC):
         # takes a dataframe and swaps branches init_flows < 0
         self.branch_direction_swaps(df)
 
-        new_flows = self.cut_lines_and_recomputes_flows(line_to_cut)
-        # print("new simulated flows = ", new_flows)
+        new_flows = np.asarray(self.cut_lines_and_recomputes_flows(line_to_cut), dtype=float)
+        swapped = df["swapped"].to_numpy()
 
         # here we multiply by (-1) new flows that are reversed
-        n_flows = []
-        for f, swapped in zip(new_flows, df["swapped"]):
-            if swapped:
-                n_flows.append(f * -1)
-            else:
-                n_flows.append(f)
+        new_flows_signed = np.where(swapped, -new_flows, new_flows)
+        df["new_flows"] = new_flows_signed
 
-        df["new_flows"] = n_flows
+        init_flows = df["init_flows"].to_numpy(dtype=float)
+        abs_new = np.abs(new_flows_signed)
+        abs_init = np.abs(init_flows)
 
         # if new_flows < 0, and abs(new) > abs(init) then True (we invert edge direction) else False
-        new_flows_swapped = []
-
-        for i, row in df.iterrows():
-            # if newf < 0. and fabs(new_flows) > fabs(initf):
-            if row["new_flows"] < 0 and fabs(row["new_flows"]) > fabs(row["init_flows"]):
-                new_flows_swapped.append(True)
-            else:
-                new_flows_swapped.append(False)
-
+        new_flows_swapped = (new_flows_signed < 0) & (abs_new > abs_init)
         df["new_flows_swapped"] = new_flows_swapped
 
-        delta_flo = []
-
         # now we add delta flows
-        # report=abs(new_flows) - abs(init_flows) si le flux n'a pas change de direction
-        # Si le flux a changé de direction, il y a 2 cas:
-        # soit le nouveau flux est plus faible et dans ce cas, le report est négatif (on a déchargé la ligne) et le
-        # report = -(abs(new_flows) + abs(init_flows))
-        # sinon le report est positif et le report =
-        # report = abs(new_flows) + abs(init_flows)
-        for i, row in df.iterrows():
-            if row["new_flows_swapped"]:
-                delta_flo.append(fabs(row["new_flows"]) + fabs(row["init_flows"]))
-                # here we swap origin and ext
-                idx_or = row["idx_or"]
-                df.at[i, "idx_or"] = row["idx_ex"]
-                df.at[i, "idx_ex"] = idx_or
-                df.at[i, "init_flows"] = fabs(row["init_flows"])
-                # print(f"row #{i}, swapped idxor and idxer")
-            elif (np.sign(row["new_flows"])!=np.sign(row["init_flows"])) and (row["new_flows"]!=0) and (row["init_flows"]!=0):#sign of 0 value is 0...
-                delta_flo.append(-(fabs(row["new_flows"]) + fabs(row["init_flows"])))#negative flow dispacth in that case
-            else:
-                delta_flo.append(fabs(row["new_flows"]) - fabs(row["init_flows"]))
-
-        # delta_flows = self.df["new_flows"].abs() - self.df["init_flows"].abs()
+        # report = abs(new) - abs(init) if the flow did not change direction.
+        # If it did there are two cases:
+        #   * new_flows_swapped -> the edge is inverted: report = abs(new) + abs(init)
+        #   * opposite signs (both non-zero) -> discharged: report = -(abs(new) + abs(init))
+        opposite_sign = (
+            (np.sign(new_flows_signed) != np.sign(init_flows))
+            & (new_flows_signed != 0)
+            & (init_flows != 0)
+        )
+        # ``elif`` semantics: opposite_sign only applies where new_flows_swapped is False.
+        discharged = opposite_sign & ~new_flows_swapped
+        delta_flo = np.where(
+            new_flows_swapped, abs_new + abs_init,
+            np.where(discharged, -(abs_new + abs_init), abs_new - abs_init),
+        )
         df["delta_flows"] = delta_flo
 
-        # small modification test to have multiple components
-        # self.df.set_value(0, "delta_flows", -32.)
-        # self.df.set_value(4, "delta_flows", -42.)
-        # self.df.set_value(5, "delta_flows", -22.)
+        # swap origin/extremity (and abs the init flow) on inverted edges
+        idx_or = df["idx_or"].to_numpy()
+        idx_ex = df["idx_ex"].to_numpy()
+        df["idx_or"] = np.where(new_flows_swapped, idx_ex, idx_or)
+        df["idx_ex"] = np.where(new_flows_swapped, idx_or, idx_ex)
+        df["init_flows"] = np.where(new_flows_swapped, abs_init, init_flows)
 
-        # DO NOT USE SET_VALUE ANYMORE, USE DF.AT INSTEAD
-        # df.at[5, "delta_flows"] = -22.
-
-        # now we identify gray edges
-        gray_edges = []
-        ltc_report = df["delta_flows"].abs()[line_to_cut[0]]#pd.DataFrame.max(df["delta_flows"].abs())
-        # print("max = ", max_report)
+        # now we identify gray edges (below-significance redispatch).
+        # ``.iloc`` makes the positional (row == line id) access explicit.
+        ltc_report = float(df["delta_flows"].abs().iloc[line_to_cut[0]])
         max_overload = ltc_report * float(self.param_options["ThresholdReportOfLine"])
-        # print("max overload = ", max_overload)
-        for edge_value in df["delta_flows"]:
-            if fabs(edge_value) < max_overload:
-                gray_edges.append(True)
-            else:
-                gray_edges.append(False)
-        # print("gray edges = ", gray_edges)
-        df["gray_edges"] = gray_edges
+        df["gray_edges"] = df["delta_flows"].abs().to_numpy() < max_overload
 
         if getattr(self, "debug", False):
             logger.debug("==== After gray_edges added IN FUNCTION CREATE DF ====")
@@ -300,27 +280,21 @@ class Simulation(ABC):
 
     @staticmethod
     def branch_direction_swaps(df: pd.DataFrame) -> None:
-        """we parse self.df and invert branches init_flows < 0"""
-        swapped = []
-        for i, row in df.iterrows():
-            # print("i {} row {}".format(i, row))
-            # a = row["delta_flows"]
-            # b = row["final_delta_flows"]
-            # if np.sign(a) != np.sign(b):
+        """Invert branches whose ``init_flows`` is negative (draw them forward).
 
-            a = row["init_flows"]
-            if a < 0 and a != 0.:
-                # here we swap origin and ext
-                idx_or = row["idx_or"]
-                df.at[i, "idx_or"] = row["idx_ex"]
-                df.at[i, "idx_ex"] = idx_or
-                df.at[i, "init_flows"] = fabs(row["init_flows"])
-                # print(f"row #{i}, swapped idxor and idxer")
-                swapped.append(True)
-            else:
-                swapped.append(False)
+        Vectorised replacement for the former ``iterrows`` loop; identical
+        results. Rows with ``init_flows < 0`` get their origin/extremity swapped,
+        their ``init_flows`` made positive, and ``swapped=True``.
+        """
+        init = df["init_flows"].to_numpy(dtype=float)
+        swap_mask = (init < 0) & (init != 0.0)
 
-        df["swapped"] = swapped
+        idx_or = df["idx_or"].to_numpy()
+        idx_ex = df["idx_ex"].to_numpy()
+        df["idx_or"] = np.where(swap_mask, idx_ex, idx_or)
+        df["idx_ex"] = np.where(swap_mask, idx_or, idx_ex)
+        df["init_flows"] = np.where(swap_mask, np.abs(init), init)
+        df["swapped"] = swap_mask
 
     @staticmethod
     def invert_dict_keys_values(d: Dict[Any, Any]) -> Dict[Any, Any]:
