+++ alphaDeesp/tests/test_graphs_package.py
@@ -44,6 +44,14 @@ EXPECTED_PUBLIC_NAMES = frozenset({
     "OverflowGraphRenderer",
     "ConstrainedPath",
     "Structured_Overload_Distribution_Graph",
+    "edge_role_of",
+    "base_color_of",
+    "EDGE_ROLE_OVERLOAD",
+    "EDGE_ROLE_NEGATIVE",
+    "EDGE_ROLE_POSITIVE",
+    "EDGE_ROLE_INSIGNIFICANT",
+    "EDGE_ROLE_NULL_NON_RECONNECTABLE",
+    "EDGE_ROLE_UNKNOWN",
     "from_edges_get_nodes",
     "delete_color_edges",
     "nodepath_to_edgepath",
@@ -67,6 +75,14 @@ EXPECTED_SYMBOL_SUBMODULE = {
     "ConstrainedPath": "alphaDeesp.core.graphs.constrained_path",
     "Structured_Overload_Distribution_Graph":
         "alphaDeesp.core.graphs.structured_overload_graph",
+    "edge_role_of": "alphaDeesp.core.graphs.edge_roles",
+    "base_color_of": "alphaDeesp.core.graphs.edge_roles",
+    "EDGE_ROLE_OVERLOAD": "alphaDeesp.core.graphs.edge_roles",
+    "EDGE_ROLE_NEGATIVE": "alphaDeesp.core.graphs.edge_roles",
+    "EDGE_ROLE_POSITIVE": "alphaDeesp.core.graphs.edge_roles",
+    "EDGE_ROLE_INSIGNIFICANT": "alphaDeesp.core.graphs.edge_roles",
+    "EDGE_ROLE_NULL_NON_RECONNECTABLE": "alphaDeesp.core.graphs.edge_roles",
+    "EDGE_ROLE_UNKNOWN": "alphaDeesp.core.graphs.edge_roles",
     "from_edges_get_nodes": "alphaDeesp.core.graphs.graph_utils",
     "delete_color_edges": "alphaDeesp.core.graphs.graph_utils",
     "nodepath_to_edgepath": "alphaDeesp.core.graphs.graph_utils",
@@ -82,6 +98,7 @@ EXPECTED_SYMBOL_SUBMODULE = {
 
 EXPECTED_SUBMODULES = frozenset({
     "constants",
+    "edge_roles",
     "graph_utils",
     "null_flow",
     "shortest_paths",
@@ -324,6 +341,31 @@ class TestStructuredOverloadDistributionGraph:
         lines, nodes = sg.get_dispatch_edges_nodes()
         assert set(lines) == {"line_AX", "line_XD"}
         assert set(nodes) == {"A", "X", "D"}
+
+    def test_derived_colour_views_are_cached(self):
+        """Lazy cached properties return the same object on repeated access."""
+        sg = Structured_Overload_Distribution_Graph(_make_structured_overload_input())
+        assert sg.g_only_red_components is sg.g_only_red_components
+        assert sg.g_only_blue_components is sg.g_only_blue_components
+        assert sg.g_without_constrained_edge is sg.g_without_constrained_edge
+
+    def test_lazy_results_are_order_independent(self):
+        """Accessing hubs before loops (or vice-versa) yields identical results —
+        the ``red_loops`` cache is keyed on the seed hubs, not the detected ones."""
+        sg_a = Structured_Overload_Distribution_Graph(_make_structured_overload_input())
+        loops_first = [list(p) for p in sg_a.red_loops["Path"].tolist()]
+        hubs_after = set(sg_a.get_hubs())
+
+        sg_b = Structured_Overload_Distribution_Graph(_make_structured_overload_input())
+        hubs_first = set(sg_b.get_hubs())
+        loops_after = [list(p) for p in sg_b.red_loops["Path"].tolist()]
+
+        assert loops_first == loops_after
+        assert hubs_after == hubs_first == {"A", "D"}
+
+    def test_get_loops_returns_the_cached_seed_red_loops(self):
+        sg = Structured_Overload_Distribution_Graph(_make_structured_overload_input())
+        assert sg.get_loops() is sg.red_loops
 
 
 def _make_loopless_overload_input():
