+++ setup.py
@@ -55,6 +55,7 @@ setup(name='ExpertOp4Grid',
       packages=setuptools.find_packages(),
       extras_require=pkgs["extras"],
       include_package_data=True,
+      package_data={"alphaDeesp.core.interactive_html": ["assets/*"]},
       install_requires=pkgs["required"],
       zip_safe=False,
       entry_points={'console_scripts': ['expertop4grid=alphaDeesp.main:main']},
